Trouble in Paradise hashing & integer value changes

+++ ReaderUtils.h
@@ -4,7 +4,7 @@
 /// </summary>
 /// <param name="val"></param>
 /// <returns><paramref name="val"/>, opposite of the current endian.</returns>
-static int flipEndian(int val) {
+static int32_t flipEndian(int32_t val) {
 	return ((0xFF000000 & val) >> 24) | ((0x00FF0000 & val) >> 8) | ((0x0000FF00 & val) << 8) | ((0x000000FF & val) << 24);
 }
 
@@ -13,7 +13,7 @@ static int flipEndian(int val) {
 /// </summary>
 /// <param name="val"></param>
 /// <returns><paramref name="val"/>, opposite of the current endian.</returns>
-static unsigned int flipEndian(unsigned int val) {
+static uint32_t flipEndian(uint32_t val) {
 	return ((0xFF000000 & val) >> 24) | ((0x00FF0000 & val) >> 8) | ((0x0000FF00 & val) << 8) | ((0x000000FF & val) << 24);
 }
 
@@ -22,6 +22,6 @@ static unsigned int flipEndian(unsigned int val) {
 /// </summary>
 /// <param name="val"></param>
 /// <returns><paramref name="val"/>, opposite of the current endian.</returns>
-static unsigned long long flipEndian(unsigned long long val) {
+static uint64_t flipEndian(uint64_t val) {
 	return ((0xFF00000000000000 & val) >> 56) | ((0x00FF000000000000 & val) >> 40) | ((0x0000FF0000000000 & val) >> 24) | ((0x000000FF00000000 & val) >> 8) | ((0x00000000ff000000 & val) << 8) | ((0x0000000000ff0000 & val) << 24) | ((0x000000000000ff00 & val) << 40) | ((0x00000000000000ff & val) << 56);
 }
